coin guide rafector, add worker name section, add view dashboard section, copy
