+++ src/App/App.scss
@@ -16,6 +16,7 @@ body {
   --success: #15cd72;
   --success-shadow: rgba(21, 205, 114, 0.5);
   --danger: #ed4f32;
+  --error: #ed4f32;
   --danger-shadow: rgba(237, 79, 50, 0.5);
   --warning: #edb431;
   --warning-shadow: rgba(237, 180, 49, 0.5);
