+++ src/components/InfoBox.tsx
@@ -1,20 +1,23 @@
 import React from 'react';
-import styled from 'styled-components/macro';
+import styled, { UIVariant } from 'styled-components/macro';
 
-const InfoBoxWrapper = styled.div<{ variant?: 'error' }>`
+const InfoBoxWrapper = styled.div<{ variant?: 'error' | UIVariant }>`
   padding: 1rem;
   border-radius: 5px;
   background: var(--primary);
   color: var(--text-on-bg);
-  ${(p) =>
-    p.variant === 'error' &&
-    `
-    background: var(--danger);
-  `}
+
+  ${(p) => {
+    if (p.variant) {
+      return `
+      background-color: var(--${p.variant});
+      `;
+    }
+  }}
 `;
 
 export const InfoBox: React.FC<{
-  variant: 'error';
+  variant: 'error' | UIVariant;
   children: React.ReactNode;
 }> = ({ variant, children }) => (
   <InfoBoxWrapper variant={variant}>{children}</InfoBoxWrapper>
