+++ src/pages/GetStarted/CoinGuide.tsx
@@ -1,83 +1,14 @@
 import React from 'react';
-import { FaDownload } from 'react-icons/fa';
 import { Redirect, useRouteMatch } from 'react-router-dom';
-import { Button } from 'src/components/Button';
-import { Img } from 'src/components/Img';
-import { Card, CardBody } from 'src/components/layout/Card';
 import { Page } from 'src/components/layout/Page';
-import { LinkOut } from 'src/components/LinkOut';
-import { Highlight } from 'src/components/Typo/Typo';
-import { getOsLogoUrl } from 'src/utils/staticImage.utils';
-import styled from 'styled-components/macro';
+import { Spacer } from 'src/components/layout/Spacer';
 
 import { mineableCoins } from './mineableCoinList';
-import { MinerCommand } from './MinerCommand';
-import { PingTest } from './PingTest';
-import { SetWallet } from './SetWallet';
-
-const MinerHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  @media screen and (max-width: 600px) {
-    display: block;
-  }
-`;
-
-const Fee: React.FC<{ fee: [number] | [number, number] }> = ({ fee }) => {
-  if (fee.length === 1) {
-    return <>{fee[0]}%</>;
-  }
-  if (fee.length === 2) {
-    return (
-      <>
-        {fee[0]}-{fee[1]}%
-      </>
-    );
-  }
-  return null;
-};
-
-const PlatformSticker = styled.span<{ gpu: 'NVIDIA' | 'AMD' | string }>`
-  padding: 0.25rem 0.5rem;
-  border-radius: 5px;
-  display: inline-block;
-  margin-left: 0.5rem;
-  font-family: 'Roboto Mono';
-  font-size: 0.875rem;
-  ${(p) => {
-    switch (p.gpu) {
-      case 'AMD':
-        return `
-          background: #EF0707;
-          color: var(--text-on-bg);
-      `;
-      case 'NVIDIA':
-        return `
-          background: #7ab547;
-          color: var(--text-on-bg);
-      `;
-    }
-  }}
-`;
-
-const SoftwareWrapper = styled.div`
-  & > * {
-    margin-bottom: 1rem;
-  }
-`;
-
-const OsLogo = styled(Img)`
-  height: 30px;
-`;
-const OsContainer = styled.div`
-  display: flex;
-  justify-content: flex-end;
-  margin-top: 0.5rem;
-  align-items: center;
-  & > * {
-    margin-left: 0.5rem;
-  }
-`;
+import { MinerCommandSection } from './MinerCommand.section';
+import { PingTestSection } from './PingTest.section';
+import { SetWalletSection } from './SetWallet.section';
+import { SetWorkerNameSection } from './SetWorkerName.section';
+import { ViewDashboardSection } from './ViewDashboard.section';
 
 export const MineableCoinGuidePage: React.FC = () => {
   const {
@@ -102,54 +33,15 @@ export const MineableCoinGuidePage: React.FC = () => {
   return (
     <Page>
       <h1>{mineableCoin.name} mining</h1>
-      <SetWallet />
-      {mineableCoin && <PingTest data={mineableCoin.regions} />}
-      <h2>
-        <Highlight>#3</Highlight> Choose your mining software
-      </h2>
-      <SoftwareWrapper>
-        {mineableCoinConfig?.miners.map((miner) => (
-          <Card key={miner.key}>
-            <CardBody>
-              <MinerHeader>
-                <div>
-                  <h3>
-                    {miner.title}{' '}
-                    {miner.compatibleGpus.map((gpu) => (
-                      <PlatformSticker key={gpu} gpu={gpu.toUpperCase()}>
-                        {gpu}
-                      </PlatformSticker>
-                    ))}
-                  </h3>
-                  <p>{miner.description}</p>
-                  <p>
-                    <strong>Fee: </strong> <Fee fee={miner.fee} />
-                  </p>
-                </div>
-                <OsContainer>
-                  {miner.os.map((osItem) => (
-                    <OsLogo
-                      src={getOsLogoUrl(osItem)}
-                      alt={`${osItem} logo`}
-                      key={osItem}
-                    />
-                  ))}
-                  <Button
-                    shape="square"
-                    variant="primary"
-                    size="sm"
-                    as={LinkOut}
-                    href={miner.downloadLink}
-                  >
-                    <FaDownload />
-                  </Button>
-                </OsContainer>
-              </MinerHeader>
-            </CardBody>
-            <MinerCommand command={miner.cmd} />
-          </Card>
-        ))}
-      </SoftwareWrapper>
+      <SetWalletSection />
+      <Spacer size="xl" />
+      <PingTestSection data={mineableCoin.regions} />
+      <Spacer size="xl" />
+      <SetWorkerNameSection />
+      <Spacer size="xl" />
+      <MinerCommandSection data={mineableCoinConfig.miners} />
+      <Spacer size="xl" />
+      <ViewDashboardSection ticker={ticker} />
     </Page>
   );
 };
